Complete etch-a-sketch game
Added erase function with erase button that toggles and makes
the pen to be permanently white.

Adjusted code to be able to reupdate what the current cells are so
that the cells can be colored after a new grid is made.

Created scaling function for the new grid maker.

Established ui for etch-a-sketch in style.css.